refactor: replace sqlx with libsql — rename ldb to db, delete old modules
Delete the sqlx-based db, page, and domain_signup modules. Promote
the libsql-based ldb module to become the new db module. Update
Cargo.toml features so session and job have their own feature flags
(both imply db). Update health check, config, lib.rs re-exports,
integration test feature gates, and all doc references from ldb to db.

Session, job, and testing modules still reference sqlx and will not
compile until ported in subsequent tasks.
