+++ src/config/modo.rs
@@ -14,14 +14,10 @@ use serde::Deserialize;
 pub struct Config {
     /// HTTP server bind address and shutdown behaviour.
     pub server: crate::server::Config,
-    /// SQLite connection pool settings. Requires the `db` feature.
+    /// libsql database settings. Requires the `db` feature.
     #[cfg(feature = "db")]
     #[serde(default)]
     pub database: crate::db::Config,
-    /// libsql database settings. Requires the `ldb` feature.
-    #[cfg(feature = "ldb")]
-    #[serde(default)]
-    pub ldb: crate::ldb::Config,
     /// Log level, format, and optional Sentry integration.
     pub tracing: crate::tracing::Config,
     /// Signed cookie secret and attributes. When absent, signed/private cookies
@@ -36,21 +32,17 @@ pub struct Config {
     /// Token-bucket rate-limiting settings.
     pub rate_limit: crate::middleware::RateLimitConfig,
     /// Session TTL, cookie name, fingerprint validation, touch interval, and
-    /// per-user session limit. Requires the `db` feature.
-    #[cfg(feature = "db")]
+    /// per-user session limit. Requires the `session` feature.
+    #[cfg(feature = "session")]
     #[serde(default)]
     pub session: crate::session::SessionConfig,
-    /// Pagination defaults (items per page, max page size). Requires the `db` feature.
-    #[cfg(feature = "db")]
-    #[serde(default)]
-    pub pagination: crate::page::PaginationConfig,
     /// HTTP client settings (timeout, retries, user agent).
     /// Requires the `http-client` feature.
     #[cfg(feature = "http-client")]
     #[serde(default)]
     pub http: crate::http::ClientConfig,
-    /// Background job queue settings. Requires the `db` feature.
-    #[cfg(feature = "db")]
+    /// Background job queue settings. Requires the `job` feature.
+    #[cfg(feature = "job")]
     #[serde(default)]
     pub job: crate::job::JobConfig,
     /// CIDR ranges of trusted reverse proxies used by [`crate::ip::ClientIpLayer`].
