+++ src/db/connect.rs
@@ -1,179 +1,87 @@
-use std::time::Duration;
+use crate::error::Result;
 
-use crate::error::{Error, Result};
+use super::config::Config;
+use super::database::Database;
 
-use super::config::SqliteConfig;
-use super::pool::{Pool, ReadPool, WritePool};
-
-/// Opens a single SQLite connection pool from `config`.
+/// Open a local libsql database, apply PRAGMAs from [`Config`], and
+/// optionally run migrations.
 ///
-/// All configured PRAGMAs (`journal_mode`, `synchronous`, `foreign_keys`,
-/// `busy_timeout`, `cache_size`, and optionally `mmap_size`, `temp_store`,
-/// `wal_autocheckpoint`) are applied on every new connection via
-/// `after_connect`.
-///
-/// The database file and any missing parent directories are created
-/// automatically when `path` is not `":memory:"`.
-///
-/// For `":memory:"` databases with `max_connections > 1`, this function logs a
-/// warning and forces `max_connections = 1` because each SQLite connection to
-/// `:memory:` gets its own isolated database instance.
+/// Creates parent directories for the database path if they do not exist.
+/// When [`Config::migrations`] is set, SQL migration files from that
+/// directory are applied via [`migrate`](super::migrate).
 ///
 /// # Errors
 ///
-/// Returns [`crate::Error::internal`] if the pool cannot be created or if the
-/// database directory cannot be created.
-pub async fn connect(config: &SqliteConfig) -> Result<Pool> {
-    let url = build_url(&config.path, false)?;
-
-    let overrides = if config.path == ":memory:" && config.max_connections > 1 {
-        tracing::warn!(
-            "in-memory database: forcing max_connections=1 (each connection gets a separate database)"
-        );
-        Some(super::config::PoolOverrides {
-            max_connections: Some(1),
-            min_connections: Some(1),
-            ..Default::default()
-        })
-    } else {
-        None
-    };
-
-    let pool = build_sqlite_pool(&url, config, overrides.as_ref()).await?;
-    Ok(Pool::new(pool))
-}
-
-/// Opens separate read and write SQLite connection pools from `config`.
-///
-/// The read pool uses `config.reader` overrides (defaulting to many
-/// connections, lower busy timeout, and large cache). The write pool uses
-/// `config.writer` overrides (defaulting to `max_connections = 1` to
-/// serialize writes).
-///
-/// All configured PRAGMAs are applied on every new connection for both pools.
-///
-/// # Errors
-///
-/// Returns [`crate::Error::internal`] if `config.path` is `":memory:"` —
-/// in-memory databases cannot be shared across separate connection pools.
-/// Use [`connect`] and wrap via [`ReadPool::new`] / [`WritePool::new`] to
-/// share a single in-memory pool in tests.
-///
-/// Also returns [`crate::Error::internal`] if either pool cannot be created.
-pub async fn connect_rw(config: &SqliteConfig) -> Result<(ReadPool, WritePool)> {
-    if config.path == ":memory:" {
-        return Err(Error::internal(
-            "read/write split is not supported for in-memory SQLite databases",
-        ));
-    }
-
-    let writer_url = build_url(&config.path, false)?;
-    let reader_url = build_url(&config.path, true)?;
-    let writer_pool = build_sqlite_pool(&writer_url, config, Some(&config.writer)).await?;
-    let reader_pool = build_sqlite_pool(&reader_url, config, Some(&config.reader)).await?;
-
-    Ok((ReadPool::new(reader_pool), WritePool::new(writer_pool)))
-}
-
-fn build_url(path: &str, read_only: bool) -> Result<String> {
-    if path == ":memory:" {
-        return Ok("sqlite::memory:".to_string());
-    }
-
-    let path = std::path::Path::new(path);
-    if let Some(parent) = path.parent()
+/// Returns an error if directory creation, database opening, PRAGMA
+/// execution, or migration fails.
+pub async fn connect(config: &Config) -> Result<Database> {
+    // Create parent directories if needed
+    if let Some(parent) = std::path::Path::new(&config.path).parent()
         && !parent.as_os_str().is_empty()
-        && !parent.exists()
     {
-        std::fs::create_dir_all(parent)
-            .map_err(|e| Error::internal(format!("failed to create database directory: {e}")))?;
+        tokio::fs::create_dir_all(parent).await.map_err(|e| {
+            crate::error::Error::internal(format!(
+                "failed to create database directory: {parent:?}"
+            ))
+            .chain(e)
+        })?;
     }
 
-    let mode = if read_only { "ro" } else { "rwc" };
-    Ok(format!("sqlite://{}?mode={mode}", path.display()))
-}
-
-async fn build_sqlite_pool(
-    url: &str,
-    config: &SqliteConfig,
-    overrides: Option<&super::config::PoolOverrides>,
-) -> Result<sqlx::SqlitePool> {
-    use sqlx::sqlite::SqlitePoolOptions;
-
-    let max_conn = overrides
-        .and_then(|o| o.max_connections)
-        .unwrap_or(config.max_connections);
-    let min_conn = overrides
-        .and_then(|o| o.min_connections)
-        .unwrap_or(config.min_connections);
-    let acquire_timeout = overrides
-        .and_then(|o| o.acquire_timeout_secs)
-        .unwrap_or(config.acquire_timeout_secs);
-    let idle_timeout = overrides
-        .and_then(|o| o.idle_timeout_secs)
-        .unwrap_or(config.idle_timeout_secs);
-    let max_lifetime = overrides
-        .and_then(|o| o.max_lifetime_secs)
-        .unwrap_or(config.max_lifetime_secs);
-    let busy_timeout = overrides
-        .and_then(|o| o.busy_timeout)
-        .unwrap_or(config.busy_timeout);
-    let cache_size = overrides
-        .and_then(|o| o.cache_size)
-        .unwrap_or(config.cache_size);
-    let mmap_size = overrides.and_then(|o| o.mmap_size).or(config.mmap_size);
-    let temp_store = overrides.and_then(|o| o.temp_store).or(config.temp_store);
-    let wal_autocheckpoint = overrides
-        .and_then(|o| o.wal_autocheckpoint)
-        .or(config.wal_autocheckpoint);
-
-    let journal_mode = config.journal_mode;
-    let synchronous = config.synchronous;
-    let foreign_keys = config.foreign_keys;
-
-    let pool = SqlitePoolOptions::new()
-        .max_connections(max_conn)
-        .min_connections(min_conn)
-        .acquire_timeout(Duration::from_secs(acquire_timeout))
-        .idle_timeout(Duration::from_secs(idle_timeout))
-        .max_lifetime(Duration::from_secs(max_lifetime))
-        .after_connect(move |conn, _meta| {
-            Box::pin(async move {
-                use sqlx::Executor;
-                conn.execute(format!("PRAGMA journal_mode = {journal_mode}").as_str())
-                    .await?;
-                conn.execute(format!("PRAGMA busy_timeout = {busy_timeout}").as_str())
-                    .await?;
-                conn.execute(format!("PRAGMA synchronous = {synchronous}").as_str())
-                    .await?;
-                conn.execute(
-                    format!(
-                        "PRAGMA foreign_keys = {}",
-                        if foreign_keys { "ON" } else { "OFF" }
-                    )
-                    .as_str(),
-                )
-                .await?;
-                conn.execute(format!("PRAGMA cache_size = {cache_size}").as_str())
-                    .await?;
-                if let Some(mmap) = mmap_size {
-                    conn.execute(format!("PRAGMA mmap_size = {mmap}").as_str())
-                        .await?;
-                }
-                if let Some(ts) = temp_store {
-                    conn.execute(format!("PRAGMA temp_store = {ts}").as_str())
-                        .await?;
-                }
-                if let Some(ac) = wal_autocheckpoint {
-                    conn.execute(format!("PRAGMA wal_autocheckpoint = {ac}").as_str())
-                        .await?;
-                }
-                Ok(())
-            })
-        })
-        .connect(url)
+    let db = libsql::Builder::new_local(&config.path)
+        .build()
         .await
-        .map_err(|e| Error::internal(format!("failed to connect to database: {e}")))?;
+        .map_err(crate::error::Error::from)?;
 
-    Ok(pool)
+    let conn = db.connect().map_err(crate::error::Error::from)?;
+
+    // Apply PRAGMAs (use query() because PRAGMAs return rows in libsql)
+    conn.query(
+        &format!("PRAGMA journal_mode={}", config.journal_mode.as_str()),
+        (),
+    )
+    .await
+    .map_err(crate::error::Error::from)?;
+
+    conn.query(
+        &format!("PRAGMA synchronous={}", config.synchronous.as_str()),
+        (),
+    )
+    .await
+    .map_err(crate::error::Error::from)?;
+
+    conn.query(&format!("PRAGMA busy_timeout={}", config.busy_timeout), ())
+        .await
+        .map_err(crate::error::Error::from)?;
+
+    conn.query(&format!("PRAGMA cache_size=-{}", config.cache_size), ())
+        .await
+        .map_err(crate::error::Error::from)?;
+
+    conn.query(&format!("PRAGMA mmap_size={}", config.mmap_size), ())
+        .await
+        .map_err(crate::error::Error::from)?;
+
+    conn.query(
+        &format!(
+            "PRAGMA foreign_keys={}",
+            if config.foreign_keys { "ON" } else { "OFF" }
+        ),
+        (),
+    )
+    .await
+    .map_err(crate::error::Error::from)?;
+
+    conn.query(
+        &format!("PRAGMA temp_store={}", config.temp_store.as_str()),
+        (),
+    )
+    .await
+    .map_err(crate::error::Error::from)?;
+
+    // Run migrations if configured
+    if let Some(ref migrations_dir) = config.migrations {
+        super::migrate::migrate(&conn, migrations_dir).await?;
+    }
+
+    Ok(Database::new(db, conn))
 }
