--- src/page/mod.rs
@@ -1,12 +1,0 @@
-mod config;
-mod cursor;
-mod offset;
-mod request;
-mod response;
-mod value;
-
-pub use config::PaginationConfig;
-pub use cursor::CursorPaginate;
-pub use offset::Paginate;
-pub use request::{CursorRequest, PageRequest};
-pub use response::{CursorPage, Page};
