+++ tests/session_meta_test.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "db")]
+#![cfg(feature = "session")]
 
 use http::HeaderMap;
 use modo::ip::extract_client_ip;
